# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dent App API",
    "description": "REST API for the Dent App quiz platform. This document was generated from the Express/MongoDB backend contained in this repository. It describes all public and administrative endpoints that the frontend relies on, including authentication flows, quiz management, questions, and quiz result tracking.",
    "version": "1.0.0",
    "contact": {
      "name": "Dent App Backend",
      "url": "https://github.com/DmitriiBank/dent-app-v2/tree/main/backend"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3555",
      "description": "Local development server"
    },
    {
      "url": "https://dent-app-v2.onrender.com",
      "description": "Production gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "fail" },
          "message": { "type": "string" }
        },
        "required": ["status", "message"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "success" },
          "token": {
            "type": "string",
            "description": "JWT access token valid for the configured expiration window."
          },
          "data": {
            "type": "object",
            "properties": {
              "safeUser": { "$ref": "#/components/schemas/User" }
            }
          }
        },
        "required": ["status", "token", "data"]
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["user", "admin"] },
          "avatar": { "type": "string", "nullable": true },
          "provider": {
            "type": "string",
            "description": "Authentication provider (\"local\" or \"google\")."
          },
          "testResults": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TestResult" }
          }
        },
        "required": ["_id", "name", "email", "role"]
      },
      "UserSignupRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password", "minLength": 8 },
          "passwordConfirm": { "type": "string", "format": "password" }
        },
        "required": ["name", "email", "password", "passwordConfirm"]
      },
...
      "Quiz": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "icon": { "type": "string", "nullable": true },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" }
          }
        },
        "required": ["_id", "title", "description"]
      },
      "QuizInput": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "icon": { "type": "string" }
        },
        "required": ["title", "description"]
      },
      "Question": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "quiz": {
            "type": "string",
            "description": "Identifier of the quiz that owns the question."
          },
          "question": { "type": "string" },
          "image": { "type": "string", "nullable": true },
          "options": {
            "type": "array",
            "minItems": 2,
            "items": { "type": "string" }
          },
          "answer": {
            "type": "integer",
            "description": "Index of the correct option in the options array."
          }
        },
        "required": ["_id", "quiz", "question", "options", "answer"]
      },
...
      "TestResult": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "quiz": { "type": "string" },
          "user": { "type": "string" },
          "points": { "type": "integer" },
          "totalQuestions": { "type": "integer" },
          "date": { "type": "string", "format": "date-time" }
        },
        "required": ["_id", "quiz", "user", "points", "totalQuestions", "date"]
      }
    },
    "parameters": {
      "PaginationPage": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1 },
        "description": "Page number for paginated responses. Defaults to 1."
      },
      "PaginationLimit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100 },
        "description": "Maximum number of items per page. Defaults to 100."
      },
      "SortParam": {
        "name": "sort",
        "in": "query",
        "schema": { "type": "string" },
        "description": "Comma separated fields to sort by, prefix with `-` for descending order."
      },
      "FieldsParam": {
        "name": "fields",
        "in": "query",
        "schema": { "type": "string" },
        "description": "Comma separated list of fields to include in the response."
      }
    }
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PaginationPage" },
          { "$ref": "#/components/parameters/PaginationLimit" },
          { "$ref": "#/components/parameters/SortParam" },
          { "$ref": "#/components/parameters/FieldsParam" }
        ],
        "responses": {
          "200": {
            "description": "List of users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "results": { "type": "integer" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/User" }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Requires admin role.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      },

      "post": {
        "tags": ["Users"],
        "summary": "Create a user (admin convenience route)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserSignupRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "403": {
            "description": "Requires admin role.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/quizzes": {
      "get": {
        "tags": ["Quizzes"],
        "summary": "List quizzes",
        "parameters": [
          { "$ref": "#/components/parameters/PaginationPage" },
          { "$ref": "#/components/parameters/PaginationLimit" },
          { "$ref": "#/components/parameters/SortParam" },
          { "$ref": "#/components/parameters/FieldsParam" }
        ],
        "responses": {
          "200": {
            "description": "Array of quizzes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "results": { "type": "integer" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Quiz" }
                    }
                  }
                }
              }
            }
          }
        }
      },

      "post": {
        "tags": ["Quizzes"],
        "summary": "Create a quiz",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QuizInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quiz created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/Quiz" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Requires admin role.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
  }
}